config: change syntax for pipe-command in key binding
The pipe command can contain almost all printable characters. In
particular, we can expect it to contain '='.

Having the pipe command as part of the key breaks the key/value
splitting.

Change it so that it is instead an optional initial part of the value,
enclosed in '[]'.

I.e. instead of

    pipe-visible:cmd=binding

we now need to write

    pipe-visible=[cmd] binding