wayland: don't assume data-device globals are added before seats
When we're handling the initial set of globals, the data-device and
primary-selection-device objects may, or may not be registered before
the seat(s).

We need to handle both cases.

When adding a seat, instantiate the data-device and
primary-selection-device **if available**.

When adding the device objects, instantiate them on **all** currently
available seats.