wayland: calculate a single, scaled DPI value
Previously, we calculated the monitors horizontal and vertical PPI
values.

For font sizing, we used the vertical PPI value. It seems however,
that using the diagonal PPI value, i.e. diagonal_px / diagonal_inch,
gives more consistent results.

This is still a PPI value, but since it is intended to be used as a
FontConfig DPI value, name it 'dpi'.